+++ tests/src/upgradability/tests/functions/set_proxy_owner.rs
@@ -1,4 +1,4 @@
-use crate::upgradeability::tests::utils::{
+use crate::upgradability::tests::utils::{
     abi_calls::{proxy_owner, set_proxy_owner},
     test_helpers::setup,
     State,